fixed nav options and support for no jquery
There are now 3 options on when calling the easyScrollDots function to allow for the existence of a sticky nav. This means that the scroll amount will now cater for the extra offset required when a fixed navigation bar is in the way.
Also, the code now no longer relies on jQuery as there is now a vanilla fallback.